First steps towards an optional keycloak dependency
