+++ shogun-boot/src/main/java/de/terrestris/shogun/boot/config/BootKeycloakConfig.java
@@ -17,7 +17,9 @@
 package de.terrestris.shogun.boot.config;
 
 import de.terrestris.shogun.config.KeycloakConfig;
+import org.springframework.boot.autoconfigure.condition.ConditionalOnExpression;
 import org.springframework.context.annotation.Configuration;
 
+@ConditionalOnExpression("${keycloak.enabled:true}")
 @Configuration
 public class BootKeycloakConfig extends KeycloakConfig { }
