+++ shogun-lib/src/main/java/de/terrestris/shogun/lib/service/GroupService.java
@@ -20,11 +20,10 @@ import de.terrestris.shogun.lib.model.Group;
 import de.terrestris.shogun.lib.model.User;
 import de.terrestris.shogun.lib.repository.GroupRepository;
 import de.terrestris.shogun.lib.repository.UserRepository;
-import de.terrestris.shogun.lib.util.KeycloakUtil;
+import de.terrestris.shogun.lib.service.security.provider.GroupProviderService;
 import lombok.extern.log4j.Log4j2;
 import org.keycloak.admin.client.resource.GroupResource;
 import org.keycloak.representations.idm.GroupRepresentation;
-import org.keycloak.representations.idm.UserRepresentation;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.jpa.domain.Specification;
 import org.springframework.security.access.prepost.PostAuthorize;
@@ -40,8 +39,11 @@ import java.util.Optional;
 @Log4j2
 public class GroupService extends BaseService<GroupRepository, Group> {
 
+//    @Autowired
+//    KeycloakUtil keycloakUtil;
+
     @Autowired
-    KeycloakUtil keycloakUtil;
+    GroupProviderService groupProviderService;
 
     @Autowired
     UserRepository userRepository;
@@ -53,7 +55,7 @@ public class GroupService extends BaseService<GroupRepository, Group> {
         List<Group> groups = (List<Group>) repository.findAll();
 
         for (Group group : groups) {
-            this.setTransientKeycloakRepresentations(group);
+            groupProviderService.setTransientRepresentations(group);
         }
 
         return groups;
@@ -66,7 +68,7 @@ public class GroupService extends BaseService<GroupRepository, Group> {
         List<Group> groups = (List<Group>) repository.findAll(specification);
 
         for (Group group : groups) {
-            this.setTransientKeycloakRepresentations(group);
+            groupProviderService.setTransientRepresentations(group);
         }
 
         return groups;
@@ -79,54 +81,10 @@ public class GroupService extends BaseService<GroupRepository, Group> {
         Optional<Group> group = repository.findById(id);
 
         if (group.isPresent()) {
-            this.setTransientKeycloakRepresentations(group.get());
+            groupProviderService.setTransientRepresentations(group.get());
         }
 
         return group;
-    }
-
-    @PostFilter("hasRole('ROLE_ADMIN') or hasPermission(filterObject, 'READ')")
-    @Transactional(readOnly = true)
-    public List<Group> findByUser(User user) {
-        List<Group> groups = new ArrayList<>();
-
-        List<GroupRepresentation> keycloakGroups = keycloakUtil.getKeycloakUserGroups(user);
-
-        for (GroupRepresentation keycloakGroup : keycloakGroups) {
-            Optional<Group> group = repository.findByKeycloakId(keycloakGroup.getId());
-            if (group.isPresent()) {
-                group.get().setKeycloakRepresentation(keycloakGroup);
-                groups.add(group.get());
-            }
-        }
-
-        return groups;
-    }
-
-    /**
-     * @deprecated Use KeycloakUtil instead
-     */
-    @Deprecated
-    public GroupRepresentation findByKeycloakId(String keycloakId) {
-        GroupResource groupResource = keycloakUtil.getGroupResource(keycloakId);
-
-        return groupResource.toRepresentation();
-    }
-
-    public List<User> getGroupMembers(String id) {
-        GroupResource groupResource = keycloakUtil.getGroupResource(id);
-        List<UserRepresentation> groupMembers = groupResource.members();
-
-        ArrayList<User> users = new ArrayList<>();
-        for (UserRepresentation groupMember : groupMembers) {
-            Optional<User> user = userRepository.findByKeycloakId(groupMember.getId());
-            if (user.isPresent()) {
-                user.get().setKeycloakRepresentation(groupMember);
-                users.add(user.get());
-            }
-        }
-
-        return users;
     }
 
     /**
@@ -152,6 +110,12 @@ public class GroupService extends BaseService<GroupRepository, Group> {
         return group;
     }
 
+    @PostFilter("hasRole('ROLE_ADMIN') or hasPermission(filterObject, 'READ')")
+    @Transactional(readOnly = true)
+    public List<Group> findByUser(User user) {
+        return groupProviderService.findByUser(user);
+    }
+
     /**
      *  Delete a group from the SHOGun DB by its keycloak Id.
      *
@@ -168,22 +132,6 @@ public class GroupService extends BaseService<GroupRepository, Group> {
         }
         repository.delete(group);
         log.info("Group with keycloak id {} was deleted in Keycloak and was therefore deleted in SHOGun DB, too.", keycloakGroupId);
-    }
-
-    private Group setTransientKeycloakRepresentations(Group group) {
-        GroupResource groupResource = keycloakUtil.getGroupResource(group);
-
-        try {
-            GroupRepresentation groupRepresentation = groupResource.toRepresentation();
-            group.setKeycloakRepresentation(groupRepresentation);
-        } catch (Exception e) {
-            log.warn("Could not get the GroupRepresentation for group with SHOGun ID {} and " +
-                    "Keycloak ID {}. This may happen if the group is not available in Keycloak.",
-                    group.getId(), group.getKeycloakId());
-            log.trace("Full stack trace: ", e);
-        }
-
-        return group;
     }
 
 }
