+++ shogun-lib/src/main/java/de/terrestris/shogun/lib/service/UserService.java
@@ -20,11 +20,9 @@ import de.terrestris.shogun.lib.enumeration.PermissionCollectionType;
 import de.terrestris.shogun.lib.event.OnRegistrationConfirmedEvent;
 import de.terrestris.shogun.lib.model.User;
 import de.terrestris.shogun.lib.repository.UserRepository;
-import de.terrestris.shogun.lib.util.KeycloakUtil;
+import de.terrestris.shogun.lib.service.security.provider.UserProviderService;
 import lombok.extern.log4j.Log4j2;
-import org.keycloak.admin.client.resource.UserResource;
 import org.keycloak.representations.idm.GroupRepresentation;
-import org.keycloak.representations.idm.UserRepresentation;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.ApplicationEventPublisher;
 import org.springframework.data.jpa.domain.Specification;
@@ -41,13 +39,7 @@ import java.util.Optional;
 public class UserService extends BaseService<UserRepository, User> {
 
     @Autowired
-    KeycloakUtil keycloakUtil;
-
-    @Autowired
-    GroupService groupService;
-
-    @Autowired
-    ApplicationEventPublisher eventPublisher;
+    UserProviderService userProviderService;
 
     @PostFilter("hasRole('ROLE_ADMIN') or hasPermission(filterObject, 'READ')")
     @Transactional(readOnly = true)
@@ -56,7 +48,7 @@ public class UserService extends BaseService<UserRepository, User> {
         List<User> users = repository.findAll();
 
         for (User user : users) {
-            this.setTransientKeycloakRepresentations(user);
+            userProviderService.setTransientRepresentations(user);
         }
 
         return users;
@@ -69,7 +61,7 @@ public class UserService extends BaseService<UserRepository, User> {
         List<User> users = (List<User>) repository.findAll(specification);
 
         for (User user : users) {
-            this.setTransientKeycloakRepresentations(user);
+            userProviderService.setTransientRepresentations(user);
         }
 
         return users;
@@ -82,7 +74,7 @@ public class UserService extends BaseService<UserRepository, User> {
         Optional<User> user = repository.findById(id);
 
         if (user.isPresent()) {
-            this.setTransientKeycloakRepresentations(user.get());
+            userProviderService.setTransientRepresentations(user.get());
         }
 
         return user;
@@ -94,51 +86,9 @@ public class UserService extends BaseService<UserRepository, User> {
         Optional<User> user = repository.findByKeycloakId(keycloakId);
 
         if (user.isPresent()) {
-            this.setTransientKeycloakRepresentations(user.get());
+            userProviderService.setTransientRepresentations(user.get());
         }
 
-        return user;
-    }
-
-    /**
-     * Finds a User by the passed keycloak ID. If it does not exists in the SHOGun DB it gets created.
-     *
-     * The groups of the user are also checked and created if needed.
-     *
-     * @param keycloakUserId
-     * @return
-     */
-//    @PreAuthorize("hasRole('ROLE_ADMIN') or hasPermission(#keycloakUserId, 'CREATE')")
-    @Transactional
-    public User findOrCreateByKeyCloakId(String keycloakUserId) {
-        Optional<User> userOptional = repository.findByKeycloakId(keycloakUserId);
-        User user = userOptional.orElse(null);
-
-        // User is not yet it SHOGun DB
-        if (user == null) {
-            user = new User(keycloakUserId, null, null, null);
-            repository.save(user);
-
-            // If the user doesn't exist, we assume it's the first login after registration.
-            eventPublisher.publishEvent(new OnRegistrationConfirmedEvent(user));
-
-            // Add admin instance permissions for the user.
-            userInstancePermissionService.setPermission(user, user, PermissionCollectionType.ADMIN);
-
-            log.info("User with keycloak id {} did not yet exist in the SHOGun DB and was therefore created.", keycloakUserId);
-            this.setTransientKeycloakRepresentations(user);
-            return user;
-        }
-
-        List<GroupRepresentation> keycloakUserGroups = keycloakUtil.getKeycloakUserGroups(user);
-
-        // Add missing groups to shogun db
-        keycloakUserGroups
-            .stream()
-            .map(GroupRepresentation::getId)
-            .forEach(groupService::findOrCreateByKeycloakId);
-
-        this.setTransientKeycloakRepresentations(user);
         return user;
     }
 
@@ -159,22 +109,6 @@ public class UserService extends BaseService<UserRepository, User> {
         userInstancePermissionService.deleteAllForEntity(user);
         repository.delete(user);
         log.info("User with keycloak id {} was deleted in Keycloak and was therefore deleted in SHOGun DB, too.", keycloakUserId);
-    }
-
-    private User setTransientKeycloakRepresentations(User user) {
-        UserResource userResource = keycloakUtil.getUserResource(user);
-
-        try {
-            UserRepresentation userRepresentation = userResource.toRepresentation();
-            user.setKeycloakRepresentation(userRepresentation);
-        } catch (Exception e) {
-            log.warn("Could not get the UserRepresentation for user with SHOGun ID {} and " +
-                    "Keycloak ID {}. This may happen if the user is not available in Keycloak.",
-                    user.getId(), user.getKeycloakId());
-            log.trace("Full stack trace: ", e);
-        }
-
-        return user;
     }
 
 }
