+++ shogun-lib/src/main/java/de/terrestris/shogun/lib/service/security/provider/keycloak/KeycloakUserProviderService.java
@@ -1,0 +1,104 @@
+package de.terrestris.shogun.lib.service.security.provider.keycloak;
+
+import de.terrestris.shogun.lib.enumeration.PermissionCollectionType;
+import de.terrestris.shogun.lib.event.OnRegistrationConfirmedEvent;
+import de.terrestris.shogun.lib.model.User;
+import de.terrestris.shogun.lib.repository.UserRepository;
+import de.terrestris.shogun.lib.service.GroupService;
+import de.terrestris.shogun.lib.service.security.permission.UserInstancePermissionService;
+import de.terrestris.shogun.lib.service.security.provider.UserProviderService;
+import de.terrestris.shogun.lib.util.KeycloakUtil;
+import lombok.extern.log4j.Log4j2;
+import org.keycloak.admin.client.resource.UserResource;
+import org.keycloak.representations.idm.GroupRepresentation;
+import org.keycloak.representations.idm.UserRepresentation;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.autoconfigure.condition.ConditionalOnExpression;
+import org.springframework.context.ApplicationEventPublisher;
+import org.springframework.stereotype.Component;
+import org.springframework.transaction.annotation.Transactional;
+
+import java.util.List;
+import java.util.Optional;
+
+@ConditionalOnExpression("${keycloak.enabled:true}")
+@Log4j2
+@Component
+public class KeycloakUserProviderService implements UserProviderService {
+
+    @Autowired
+    KeycloakUtil keycloakUtil;
+
+    @Autowired
+    UserRepository userRepository;
+
+    @Autowired
+    ApplicationEventPublisher eventPublisher;
+
+    @Autowired
+    GroupService groupService;
+
+    @Autowired
+    protected UserInstancePermissionService userInstancePermissionService;
+
+    /**
+     * Finds a User by the passed keycloak ID. If it does not exists in the SHOGun DB it gets created.
+     *
+     * The groups of the user are also checked and created if needed.
+     *
+     * @param keycloakUserId
+     * @return
+     */
+//    @PreAuthorize("hasRole('ROLE_ADMIN') or hasPermission(#keycloakUserId, 'CREATE')")
+    @Transactional
+    public User findOrCreateByProviderId(String keycloakUserId) {
+        Optional<User> userOptional = userRepository.findByKeycloakId(keycloakUserId);
+        User user = userOptional.orElse(null);
+
+        // User is not yet it SHOGun DB
+        if (user == null) {
+            user = new User(keycloakUserId, null, null, null);
+            userRepository.save(user);
+
+            // If the user doesn't exist, we assume it's the first login after registration.
+            eventPublisher.publishEvent(new OnRegistrationConfirmedEvent(user));
+
+            // Add admin instance permissions for the user.
+            userInstancePermissionService.setPermission(user, user, PermissionCollectionType.ADMIN);
+
+            log.info("User with keycloak id {} did not yet exist in the SHOGun DB and was therefore created.", keycloakUserId);
+
+            this.setTransientRepresentations(user);
+
+            return user;
+        }
+
+        List<GroupRepresentation> keycloakUserGroups = keycloakUtil.getKeycloakUserGroups(user);
+
+        // Add missing groups to shogun db
+        keycloakUserGroups
+            .stream()
+            .map(GroupRepresentation::getId)
+            .forEach(groupService::findOrCreateByKeycloakId);
+
+        this.setTransientRepresentations(user);
+
+        return user;
+    }
+
+    public User setTransientRepresentations(User user) {
+        UserResource userResource = keycloakUtil.getUserResource(user);
+
+        try {
+            UserRepresentation userRepresentation = userResource.toRepresentation();
+            user.setKeycloakRepresentation(userRepresentation);
+        } catch (Exception e) {
+            log.warn("Could not get the UserRepresentation for user with SHOGun ID {} and " +
+                    "Keycloak ID {}. This may happen if the user is not available in Keycloak.",
+                user.getId(), user.getKeycloakId());
+            log.trace("Full stack trace: ", e);
+        }
+
+        return user;
+    }
+}
